refactor[#523] : 병합 충돌 나던 부분 수정
